add unit test for generating a nodal mesh
